new Hybrid Control

+++ orangecontrib/shadow/util/shadow_util.py
@@ -294,8 +294,8 @@ try:
                 super(ShadowPlot.InfoBoxWidget, self).__init__()
 
                 info_box_inner=gui.widgetBox(self, "Info")
-                info_box_inner.setFixedHeight(518*y_scale_factor)
-                info_box_inner.setFixedWidth(230*x_scale_factor)
+                info_box_inner.setFixedHeight(int(518*y_scale_factor))
+                info_box_inner.setFixedWidth(int(230*x_scale_factor))
 
                 self.flux_box = gui.widgetBox(info_box_inner, "", addSpace=False, orientation="horizontal")
                 self.flux     = gui.lineEdit(self.flux_box, self, "flux_field", "\u03a6 [ph/s/0.1%BW]", tooltip="Flux", labelWidth=115, valueType=str, orientation="horizontal")
@@ -502,8 +502,8 @@ try:
                 self.plot_canvas = gui.plotWindow(roi=False, control=False, position=True, logScale=True, fit=True)
                 self.plot_canvas.setDefaultPlotLines(True)
                 self.plot_canvas.setActiveCurveColor(color='blue')
-                self.plot_canvas.setMinimumWidth(590*x_scale_factor)
-                self.plot_canvas.setMaximumWidth(590*x_scale_factor)
+                self.plot_canvas.setMinimumWidth(int(590*x_scale_factor))
+                self.plot_canvas.setMaximumWidth(int(590*x_scale_factor))
 
                 self.info_box = ShadowPlot.InfoBoxWidget(x_scale_factor, y_scale_factor, is_2d=False)
 
@@ -512,8 +512,8 @@ try:
                 layout.addWidget(   self.info_box, 0, 1, 1, 1)
                 layout.addWidget(self.plot_canvas, 0, 0, 1, 1)
 
-                layout.setColumnMinimumWidth(0, 600*x_scale_factor)
-                layout.setColumnMinimumWidth(1, 230*x_scale_factor)
+                layout.setColumnMinimumWidth(0, int(600*x_scale_factor))
+                layout.setColumnMinimumWidth(1, int(230*x_scale_factor))
 
                 self.setLayout(layout)
 
@@ -1304,7 +1304,6 @@ class ShadowPreProcessor:
         output_beam._beam.rays[:, 17] = output_beam._beam.rays[:, 17] * interpolated_weight_p
 
         return output_beam
-
 
 class ShadowMath:
